+++ src/Service.Tests/Configuration/ConfigurationTests.cs
@@ -4145,6 +4145,140 @@ type Planet @model(name:""PlanetAlias"") {
         }
 
         /// <summary>
+        /// Creates a hot reload scenario in which the schema file is invalid which causes
+        /// hot reload to fail, then we check that the program is still able to work
+        /// properly by validating that the DAB engine is still using the same configuration file
+        /// from before the hot reload.
+        /// </summary>
+        [TestMethod]
+        [TestCategory(TestCategory.MSSQL)]
+        public void HotReloadValidationFail()
+        {
+            // Arrange
+            string schemaName = "testSchema.json";
+            string configName = "hotreload-config.json";
+            if (File.Exists(configName))
+            {
+                File.Delete(configName);
+            }
+
+            if (File.Exists(schemaName))
+            {
+                File.Delete(schemaName);
+            }
+
+            bool initialRestEnabled = true;
+            bool updatedRestEnabled = false;
+
+            bool initialGQLEnabled = true;
+            bool updatedGQLEnabled = false;
+
+            DataSource dataSource = new(DatabaseType.MSSQL,
+                ConfigurationTests.GetConnectionStringFromEnvironmentConfig(environment: TestCategory.MSSQL), Options: null);
+
+            RuntimeConfig initialConfig = InitRuntimeConfigForHotReload(schemaName, dataSource, initialRestEnabled, initialGQLEnabled);
+
+            RuntimeConfig updatedConfig = InitRuntimeConfigForHotReload(schemaName, dataSource, updatedRestEnabled, updatedGQLEnabled);
+
+            string schemaConfig = TestHelper.GenerateInvalidSchema();
+
+            // Not using mocked filesystem so we pick up real file changes for hot reload
+            FileSystem fileSystem = new();
+            RuntimeConfigProvider configProvider = GenerateConfigFileAndConfigProvider(fileSystem, configName, initialConfig);
+            RuntimeConfig lkgRuntimeConfig = configProvider.GetConfig();
+
+            Assert.IsNotNull(lkgRuntimeConfig);
+
+            // Act
+            // Simulate an invalid change to the schema file while the config is updated to a valid state
+            fileSystem.File.WriteAllText(schemaName, schemaConfig);
+            fileSystem.File.WriteAllText(configName, updatedConfig.ToJson());
+
+            // Give ConfigFileWatcher enough time to hot reload the change
+            System.Threading.Thread.Sleep(1000);
+
+            // Validating that the config validation failed,
+            // this means that GetConfig needs to return an exception
+            try
+            {
+                configProvider.GetConfig();
+            }
+            catch (DataApiBuilderException dbException)
+            {
+                Assert.AreEqual(expected: "Failed validation of configuration file.", actual: dbException.Message);
+
+                RuntimeConfig newRuntimeConfig = configProvider.GetConfig();
+                Assert.AreEqual(expected: lkgRuntimeConfig, actual: newRuntimeConfig);
+
+                if (File.Exists(configName))
+                {
+                    File.Delete(configName);
+                }
+
+                if (File.Exists(schemaName))
+                {
+                    File.Delete(schemaName);
+                }
+
+                return;
+            }
+
+            Assert.Fail();
+        }
+
+        /// <summary>
+        /// Creates a hot reload scenario in which the updated configuration file is invalid causing
+        /// hot reload to fail as the schema can't be used by DAB, then we check that the
+        /// program is still able to work properly by showing us that it is still using the
+        /// same configuration file from before the hot reload.
+        /// </summary>
+        [TestMethod]
+        [TestCategory(TestCategory.MSSQL)]
+        public void HotReloadParsingFail()
+        {
+            // Arrange
+            string schemaName = "dab.draft.schema.json";
+            string configName = "hotreload-config.json";
+            if (File.Exists(configName))
+            {
+                File.Delete(configName);
+            }
+
+            bool initialRestEnabled = true;
+
+            bool initialGQLEnabled = true;
+
+            DataSource dataSource = new(DatabaseType.MSSQL,
+                ConfigurationTests.GetConnectionStringFromEnvironmentConfig(environment: TestCategory.MSSQL), Options: null);
+
+            RuntimeConfig initialConfig = InitRuntimeConfigForHotReload(schemaName, dataSource, initialRestEnabled, initialGQLEnabled);
+
+            string updatedConfig = TestHelper.GenerateInvalidRuntimeSection();
+
+            // Not using mocked filesystem so we pick up real file changes for hot reload
+            FileSystem fileSystem = new();
+            RuntimeConfigProvider configProvider = GenerateConfigFileAndConfigProvider(fileSystem, configName, initialConfig);
+            RuntimeConfig lkgRuntimeConfig = configProvider.GetConfig();
+
+            Assert.IsNotNull(lkgRuntimeConfig);
+
+            // Act
+            // Simulate an invalid change to the config file
+            fileSystem.File.WriteAllText(configName, updatedConfig);
+
+            // Give ConfigFileWatcher enough time to hot reload the change
+            System.Threading.Thread.Sleep(1000);
+
+            RuntimeConfig newRuntimeConfig = configProvider.GetConfig();
+            Assert.AreEqual(expected: lkgRuntimeConfig, actual: newRuntimeConfig);
+
+            if (File.Exists(configName))
+            {
+                File.Delete(configName);
+            }
+        }
+
+        /// <summary>
         /// Helper function to write custom configuration file. with minimal REST/GraphQL global settings
         /// using the supplied entities.
         /// </summary>
@@ -4594,6 +4728,53 @@ type Planet @model(name:""PlanetAlias"") {
                     Host: new(Cors: null, Authentication: authenticationOptions, Mode: HostMode.Development), Cache: cacheOptions),
                 Entities: new(entityMap)
             );
+        }
+
+        /// <summary>
+        /// Helper function that initializes a RuntimeConfig with the following
+        /// variables in order to prepare a file for hot reload
+        /// </summary>
+        /// <param name="schema"></param>
+        /// <param name="dataSource"></param>
+        /// <param name="restEnabled"></param>
+        /// <param name="graphQLEnabled"></param>
+        /// <returns></returns>
+        public static RuntimeConfig InitRuntimeConfigForHotReload(
+            string schema,
+            DataSource dataSource,
+            bool restEnabled,
+            bool graphQLEnabled)
+        {
+            RuntimeConfig runtimeConfig = new(
+                Schema: schema,
+                DataSource: dataSource,
+                Runtime: new(
+                    Rest: new(restEnabled),
+                    GraphQL: new(graphQLEnabled),
+                    Host: new(null, null, HostMode.Development)
+                ),
+                Entities: new(new Dictionary<string, Entity>())
+            );
+
+            return runtimeConfig;
+        }
+
+        /// <summary>
+        /// Helper function that generates a config file that is going to be observed
+        /// for hot reload and initialize a ConfigProvider which will be used to check
+        /// if the hot reload was validated or not.
+        /// </summary>
+        /// <param name="fileSystem"></param>
+        /// <param name="configName"></param>
+        /// <param name="runtimeConfig"></param>
+        /// <returns></returns>
+        public static RuntimeConfigProvider GenerateConfigFileAndConfigProvider(FileSystem fileSystem, string configName, RuntimeConfig runtimeConfig)
+        {
+            fileSystem.File.WriteAllText(configName, runtimeConfig.ToJson());
+            FileSystemRuntimeConfigLoader configLoader = new(fileSystem, handler: null, configName, string.Empty);
+            RuntimeConfigProvider configProvider = new(configLoader);
+
+            return configProvider;
         }
 
         /// <summary>
